feat: separated creating layout and working with input
